(fix) sha256 verification

+++ neofetch/build.sh
@@ -4,16 +4,30 @@ cd neofetch
 rm *.zip
 rm -rf neofetch-*
 
+
 # Fetch and unzip neofetch
-wget https://github.com/dylanaraps/neofetch/archive/refs/tags/7.1.0.zip .
-unzip 7.1.0.zip
-cd neofetch-7.1.0
+wget https://github.com/dylanaraps/neofetch/archive/refs/tags/7.1.0.zip
 
-# Make neofetch
-make install DESTDIR="../argon-out"
-cd ..
-rm -rf argon-out/neofetch
-zip neofetch.zip argon-out/*
+zip_hash=$(../hacky_sha 7.1.0.zip)
+echo $ARGON_ZIP_SHA
+echo $zip_hash
 
-mv neofetch.zip ../bin-pkgs/
-cd ..
+if [[ $zip_hash = $ARGON_ZIP_SHA ]]
+then
+	echo "Checksums match. Continuing build."
+	unzip 7.1.0.zip
+	cd neofetch-7.1.0
+
+	# Make neofetch
+	make install DESTDIR="../argon-out"
+	cd ..
+	rm -rf argon-out/neofetch
+	zip neofetch.zip argon-out/*
+
+	mv neofetch.zip ../bin-pkgs/
+	cd ..
+else
+	echo "Checksums DON'T match!"
+fi
+
+
